+++ p2p-service/src/p2p/mod.rs
@@ -1,5 +1,6 @@
 pub mod connection_manager;
 pub mod constants;
+pub mod device_sync;
 pub mod emitter;
 pub mod errors;
 pub mod handshake;
@@ -8,7 +9,8 @@ pub mod incoming_handler;
 pub mod logger;
 pub mod p2p_service;
 pub mod peer_connection;
-pub mod sync2;
+pub mod resource_sync;
+pub mod user_sync;
 
 pub use emitter::{P2PEvent, P2PEventEmitter};
 pub use p2p_service::P2PService;
