+++ p2p-service/src/p2p/resource_sync.rs
@@ -2,306 +2,19 @@ use crate::p2p::peer_connection::PeerConnection;
 
 use super::P2PEvent;
 use osvauld_core::models::{
-    ConnectionType, DeviceManifestComparisonResult, DeviceManifestRequestPayload,
-    DeviceNetworkSyncPayload, FirstUserExchange, LiveEditMessage, Message, ResourceSyncData,
-    ResourceUpdateMsg, User, UserManifestPayload, UserNetworkSyncPayload, UserWithDevices,
+    ConnectionType, LiveEditMessage, Message, ResourceSyncData, ResourceUpdateMsg,
 };
 use osvauld_services::{
     add_resource_sync, add_share_records, apply_updates_and_get_peer_updates,
-    create_device_network_sync_payload, create_user_network_sync_payload,
-    generate_updates_for_peer, get_device_manifest, get_my_user_devices,
-    get_resource_for_remote_addition, get_resource_state_vector, get_share_records_for_resource,
-    get_user_manifest, get_vector_clocks_for_resource, merge_share_records, merge_vector_clocks,
-    process_device_manifest_request, process_device_network_sync, process_user_manifest_request,
-    process_user_network_sync_payload, update_vector_clocks,
+    generate_updates_for_peer, get_resource_for_remote_addition, get_resource_state_vector,
+    get_share_records_for_resource, get_vector_clocks_for_resource, merge_share_records,
+    merge_vector_clocks, update_vector_clocks,
 };
 
-use tracing::{debug, error, info, instrument, Span};
+use tracing::{debug, error, info, instrument};
 
 // Helper method signatures to reduce repeated patterns
 impl PeerConnection {
-    #[instrument(skip(self), fields(
-        connection_id = %self.get_id(),
-        user_id = %self.user.id
-    ), level = "info")]
-    pub async fn start_add_device_process(&self) -> Result<(), String> {
-        info!("Starting add device process");
-        debug!("Retrieving device manifest for user");
-
-        let manifest = match get_device_manifest(&self.repo_ctx, &self.user.id).await {
-            Ok(manifest) => {
-                debug!(
-                    manifest_resource_count = manifest.resources.len(),
-                    "Device manifest retrieved successfully"
-                );
-                manifest
-            }
-            Err(e) => {
-                error!(error = %e, "Failed to get device manifest");
-                return Err(format!("Failed to get device manifest: {}", e));
-            }
-        };
-
-        match self
-            .send_message(Message::DeviceManifestRequest(manifest))
-            .await
-        {
-            Ok(_) => {
-                info!("Device manifest request sent successfully");
-                Ok(())
-            }
-            Err(e) => {
-                error!(error = %e, "Failed to send device manifest request");
-                Err(format!("Failed to send device manifest request: {}", e))
-            }
-        }
-    }
-
-    #[instrument(skip(self, payload), fields(
-        connection_id = %self.get_id(),
-        user_id = %self.user.id,
-        known_device_count = payload.known_device_ids.len(),
-        other_users_count = payload.other_users.len(),
-        folder_count = payload.folder_ids.len(),
-        resource_count = payload.resources.len()
-    ), level = "info")]
-    pub async fn handle_manifest_request(
-        &self,
-        payload: &DeviceManifestRequestPayload,
-    ) -> Result<(), String> {
-        info!("Processing device manifest request");
-        debug!(
-            "Processing manifest with {} known devices, {} other users, {} folders, {} resources",
-            payload.known_device_ids.len(),
-            payload.other_users.len(),
-            payload.folder_ids.len(),
-            payload.resources.len()
-        );
-
-        let result = match process_device_manifest_request(payload, &self.repo_ctx, &self.user.id)
-            .await
-        {
-            Ok(result) => {
-                debug!(
-                    local_missing_unknown_users = result.local_missing.unknown_users.len(),
-                    local_missing_unknown_devices_common =
-                        result.local_missing.unknown_devices_from_common_users.len(),
-                    local_missing_unknown_devices_current =
-                        result.local_missing.unknown_devices_from_current_user.len(),
-                    local_missing_unknown_resources = result.local_missing.unknown_resources.len(),
-                    local_missing_unknown_folders = result.local_missing.unknown_folders.len(),
-                    remote_missing_unknown_users = result.remote_missing.unknown_users.len(),
-                    remote_missing_unknown_devices_common = result
-                        .remote_missing
-                        .unknown_devices_from_common_users
-                        .len(),
-                    remote_missing_unknown_devices_current = result
-                        .remote_missing
-                        .unknown_devices_from_current_user
-                        .len(),
-                    remote_missing_unknown_resources =
-                        result.remote_missing.unknown_resources.len(),
-                    remote_missing_unknown_folders = result.remote_missing.unknown_folders.len(),
-                    resources_requiring_sync = result.resources_requiring_sync.len(),
-                    "Manifest comparison completed"
-                );
-                result
-            }
-            Err(e) => {
-                error!(error = %e, "Failed to process device manifest request");
-                return Err(format!("Failed to process device manifest request: {}", e));
-            }
-        };
-
-        match self
-            .send_message(Message::DeviceManifestResponse(result.clone()))
-            .await
-        {
-            Ok(_) => {
-                info!("Device manifest response sent successfully");
-                debug!("Setting manifest comparison result");
-                self.set_device_manifest_comparison_result(result).await;
-                Ok(())
-            }
-            Err(e) => {
-                error!(error = %e, "Failed to send device manifest response");
-                Err(format!("Failed to send device manifest response: {}", e))
-            }
-        }
-    }
-
-    #[instrument(skip(self, payload), fields(
-        connection_id = %self.get_id(),
-        local_missing_unknown_users = payload.local_missing.unknown_users.len(),
-        local_missing_unknown_devices_common = payload.local_missing.unknown_devices_from_common_users.len(),
-        local_missing_unknown_devices_current = payload.local_missing.unknown_devices_from_current_user.len(),
-        local_missing_unknown_resources = payload.local_missing.unknown_resources.len(),
-        local_missing_unknown_folders = payload.local_missing.unknown_folders.len(),
-        remote_missing_unknown_users = payload.remote_missing.unknown_users.len(),
-        remote_missing_unknown_devices_common = payload.remote_missing.unknown_devices_from_common_users.len(),
-        remote_missing_unknown_devices_current = payload.remote_missing.unknown_devices_from_current_user.len(),
-        remote_missing_unknown_resources = payload.remote_missing.unknown_resources.len(),
-        remote_missing_unknown_folders = payload.remote_missing.unknown_folders.len(),
-        resources_requiring_sync = payload.resources_requiring_sync.len()
-    ), level = "info")]
-    pub async fn handle_manifest_response(
-        &self,
-        payload: &DeviceManifestComparisonResult,
-    ) -> Result<(), String> {
-        info!("Processing device manifest response");
-        debug!(
-            "Received manifest comparison with {} sync-required resources",
-            payload.resources_requiring_sync.len()
-        );
-
-        let manifest_result = payload.inverse();
-        debug!(
-            inverted_local_missing_unknown_users =
-                manifest_result.local_missing.unknown_users.len(),
-            inverted_local_missing_unknown_devices_common = manifest_result
-                .local_missing
-                .unknown_devices_from_common_users
-                .len(),
-            inverted_local_missing_unknown_devices_current = manifest_result
-                .local_missing
-                .unknown_devices_from_current_user
-                .len(),
-            inverted_local_missing_unknown_resources =
-                manifest_result.local_missing.unknown_resources.len(),
-            inverted_local_missing_unknown_folders =
-                manifest_result.local_missing.unknown_folders.len(),
-            inverted_remote_missing_unknown_users =
-                manifest_result.remote_missing.unknown_users.len(),
-            inverted_remote_missing_unknown_devices_common = manifest_result
-                .remote_missing
-                .unknown_devices_from_common_users
-                .len(),
-            inverted_remote_missing_unknown_devices_current = manifest_result
-                .remote_missing
-                .unknown_devices_from_current_user
-                .len(),
-            inverted_remote_missing_unknown_resources =
-                manifest_result.remote_missing.unknown_resources.len(),
-            inverted_remote_missing_unknown_folders =
-                manifest_result.remote_missing.unknown_folders.len(),
-            "Manifest result inverted for local perspective"
-        );
-
-        self.set_device_manifest_comparison_result(manifest_result)
-            .await;
-
-        match self.send_message(Message::DeviceManifestAck).await {
-            Ok(_) => {
-                info!("Device manifest acknowledgment sent successfully");
-            }
-            Err(e) => {
-                error!(error = %e, "Failed to send device manifest ack");
-                return Err(format!("Failed to send device manifest ack: {}", e));
-            }
-        }
-
-        let manifest = self.get_device_manifest_result().await?;
-        let device_network_payload = match create_device_network_sync_payload(
-            &manifest.remote_missing,
-            &self.repo_ctx,
-        )
-        .await
-        {
-            Ok(payload) => {
-                debug!("Network sync payload created successfully");
-                payload
-            }
-            Err(e) => {
-                error!(error = %e, "Failed to create network sync payload");
-                return Err(format!("Failed to create network sync payload: {}", e));
-            }
-        };
-
-        match self
-            .send_message(Message::DeviceNetworkSync(device_network_payload))
-            .await
-        {
-            Ok(_) => {
-                info!("Device network sync message sent successfully");
-                Ok(())
-            }
-            Err(e) => {
-                error!(error = %e, "Failed to send device network sync");
-                Err(format!("Failed to send device network sync: {}", e))
-            }
-        }
-    }
-
-    #[instrument(skip(self), fields(connection_id = %self.get_id()), level = "info")]
-    pub async fn handle_manifest_ack(&self) -> Result<(), String> {
-        info!("Processing device manifest acknowledgment");
-
-        let manifest = self.get_device_manifest_result().await?;
-
-        let device_network_payload = match create_device_network_sync_payload(
-            &manifest.remote_missing,
-            &self.repo_ctx,
-        )
-        .await
-        {
-            Ok(payload) => {
-                debug!("Network sync payload created successfully for ack response");
-                payload
-            }
-            Err(e) => {
-                error!(error = %e, "Failed to create network sync payload for ack");
-                return Err(format!("Failed to create network sync payload: {}", e));
-            }
-        };
-
-        self.send_message(Message::DeviceNetworkSync(device_network_payload))
-            .await?;
-        Ok(())
-    }
-
-    #[instrument(skip(self, payload), fields(
-        connection_id = %self.get_id(),
-        unknown_users_with_devices_count = payload.unknown_users_with_devices.len(),
-        unknown_devices_common_users_count = payload.unknown_devices_from_common_users.len(),
-        unknown_devices_current_user_count = payload.unknown_devices_from_current_user.len(),
-        unknown_folders_count = payload.unknown_folders.len()
-    ), level = "info")]
-    pub async fn handle_device_network_sync(
-        &self,
-        payload: &mut DeviceNetworkSyncPayload,
-    ) -> Result<(), String> {
-        info!("Processing device network sync payload");
-        debug!(
-            "Processing network sync with {} users with devices, {} devices from common users, {} devices from current user, {} folders",
-            payload.unknown_users_with_devices.len(),
-            payload.unknown_devices_from_common_users.len(),
-            payload.unknown_devices_from_current_user.len(),
-            payload.unknown_folders.len()
-        );
-
-        match process_device_network_sync(payload, &self.repo_ctx).await {
-            Ok(_) => {
-                info!("Network sync processed successfully");
-                debug!("Sending network sync acknowledgment");
-            }
-            Err(e) => {
-                error!(error = %e, "Failed to process network sync");
-                return Err(format!("Failed to process network sync: {}", e));
-            }
-        }
-
-        match self.send_message(Message::DeviceNetworkSyncAck).await {
-            Ok(_) => {
-                info!("Device network sync acknowledgment sent successfully");
-                Ok(())
-            }
-            Err(e) => {
-                error!(error = %e, "Failed to send device network sync ack");
-                Err(format!("Failed to send device network sync ack: {}", e))
-            }
-        }
-    }
-
     #[instrument(skip(self), fields(
     connection_id = %self.get_id(),
     device_id = %self.device.id,
@@ -309,9 +22,12 @@ impl PeerConnection {
 ), level = "info")]
     pub async fn send_resources(&self) -> Result<(), String> {
         info!("Starting resource transmission process");
-
+        let connection_type = match self.connection_type.clone() {
+            Some(ct) => ct,
+            None => return Err("connection type is none".to_string()),
+        };
         // Collect resource IDs to avoid borrowing issues
-        let resource_ids: Vec<String> = match self.connection_type {
+        let resource_ids: Vec<String> = match connection_type {
             ConnectionType::Device => {
                 let manifest = self.get_device_manifest_result().await?;
                 debug!(
@@ -424,7 +140,11 @@ impl PeerConnection {
     ) -> Result<(), String> {
         info!("Processing resource addition request");
         debug!("Adding resource to local repository");
-        match add_resource_sync(payload, &self.repo_ctx, &self.connection_type).await {
+        let connection_type = match self.connection_type.clone() {
+            Some(ct) => ct,
+            None => return Err("connection type is none".to_string()),
+        };
+        match add_resource_sync(payload, &self.repo_ctx, &connection_type).await {
             Ok(_) => {
                 info!(
                     resource_id = %payload.resource.id,
@@ -443,7 +163,11 @@ impl PeerConnection {
                 return Err(format!("Failed to add resource sync: {}", e));
             }
         }
-        let is_empty = match self.connection_type {
+        let connection_type = match self.connection_type.clone() {
+            Some(ct) => ct,
+            None => return Err("connection type is none".to_string()),
+        };
+        let is_empty = match connection_type {
             ConnectionType::Device => {
                 self.remove_device_local_missing_resource(&payload.resource.id)
                     .await
@@ -469,7 +193,11 @@ impl PeerConnection {
         if self.is_initiator {
             debug!("Processing as initiator - checking for resources requiring sync");
 
-            let resource_ids: Vec<String> = match self.connection_type {
+            let connection_type = match self.connection_type.clone() {
+                Some(ct) => ct,
+                None => return Err("connection type is none".to_string()),
+            };
+            let resource_ids: Vec<String> = match connection_type {
                 ConnectionType::Device => {
                     let manifest = self.get_device_manifest_result().await?;
                     manifest.resources_requiring_sync.iter().cloned().collect()
@@ -880,118 +608,6 @@ impl PeerConnection {
         Ok(())
     }
 
-    pub async fn send_first_user_connection_payload(&self, is_request: bool) -> Result<(), String> {
-        // Only the initiator sends the UserConnection message
-        let user = self.get_local_user().await?;
-        let devices = get_my_user_devices(&user.id, &self.repo_ctx).await?;
-        let message = if is_request {
-            FirstUserExchange::Request(UserWithDevices { user, devices })
-        } else {
-            FirstUserExchange::Response(UserWithDevices { user, devices })
-        };
-        self.send_message(Message::FirstUserConnection(message))
-            .await?;
-        Ok(())
-    }
-
-    pub async fn process_first_connection_exchange(
-        &self,
-        payload: &mut FirstUserExchange,
-    ) -> Result<(), String> {
-        match payload {
-            FirstUserExchange::Request(user_with_devices) => {
-                user_with_devices.user.owner = false;
-                user_with_devices.user.first_sync = true;
-                self.repo_ctx
-                    .user_repo
-                    .add_users_with_devices_bulk(&[user_with_devices.clone()])
-                    .await
-                    .map_err(|e| e.to_string())?;
-                self.send_first_user_connection_payload(false).await?;
-            }
-            FirstUserExchange::Response(user_with_devices) => {
-                user_with_devices.user.owner = false;
-                user_with_devices.user.first_sync = true;
-                self.repo_ctx
-                    .user_repo
-                    .add_users_with_devices_bulk(&[user_with_devices.clone()])
-                    .await
-                    .map_err(|e| e.to_string())?;
-                let user_manifest = get_user_manifest(&self.repo_ctx, &self.user.id).await?;
-
-                self.send_message(Message::UserManifestPayload(UserManifestPayload::Request(
-                    user_manifest,
-                )))
-                .await?;
-            }
-        }
-        Ok(())
-    }
-
-    pub async fn start_user_network_sync(&self) -> Result<(), String> {
-        if self.is_initiator {
-            let user_manifest = get_user_manifest(&self.repo_ctx, &self.user.id).await?;
-            self.send_message(Message::UserManifestPayload(UserManifestPayload::Request(
-                user_manifest,
-            )))
-            .await?;
-        }
-        Ok(())
-    }
-
-    pub async fn process_user_manifest_payload(
-        &self,
-        payload: &UserManifestPayload,
-    ) -> Result<(), String> {
-        let user = self.get_local_user().await?;
-        match payload {
-            UserManifestPayload::Request(request_payload) => {
-                let manifest_result =
-                    process_user_manifest_request(request_payload, &self.repo_ctx, &self.user.id)
-                        .await?;
-                self.set_user_manifest_comparison_result(manifest_result.clone())
-                    .await;
-                self.send_message(Message::UserManifestPayload(UserManifestPayload::Response(
-                    manifest_result,
-                )))
-                .await?;
-            }
-            UserManifestPayload::Response(manifest) => {
-                let manifest_result = manifest.inverse();
-                self.set_user_manifest_comparison_result(manifest_result)
-                    .await;
-                self.send_message(Message::UserManifestPayload(UserManifestPayload::Ack))
-                    .await?;
-            }
-            UserManifestPayload::Ack => {
-                let manifest = self.get_user_manifest_result().await?;
-                let payload =
-                    create_user_network_sync_payload(&manifest.remote_missing, &self.repo_ctx)
-                        .await?;
-                let message = Message::UserNetworkSync(payload);
-                self.send_message(message).await?;
-            }
-        }
-        Ok(())
-    }
-
-    pub async fn process_user_network_sync(
-        &self,
-        payload: &mut UserNetworkSyncPayload,
-    ) -> Result<(), String> {
-        if self.is_initiator {
-            let manifest = self.get_user_manifest_result().await?;
-            let local_payload =
-                create_user_network_sync_payload(&manifest.remote_missing, &self.repo_ctx).await?;
-            let message = Message::UserNetworkSync(local_payload);
-            self.send_message(message).await?;
-        }
-
-        process_user_network_sync_payload(payload, &self.repo_ctx).await?;
-        self.send_message(Message::UserNetworkSyncAck).await?;
-        Ok(())
-    }
-
     #[instrument(skip(self, message), fields(message_type = ?std::mem::discriminant(message)), level = "debug")]
     pub async fn handle_live_edit_flow(&self, message: &LiveEditMessage) -> Result<(), String> {
         let user = self.get_local_user().await?;
@@ -1001,8 +617,6 @@ impl PeerConnection {
                     resource_id: resource_id.clone(),
                     connection_id: self.get_id(),
                 });
-                //mark connection for live editing.
-                self.set_live_editing_active().await;
                 Ok(())
             }
             LiveEditMessage::StateVectorExchange {
@@ -1058,18 +672,9 @@ impl PeerConnection {
                 });
                 Ok(())
             }
-            LiveEditMessage::NotSameDocument => {
-                self.set_live_editing_inactive().await;
-                // self.cancel_disconnection_timer().await;
-                // self.check_for_possible_disconnection().await?;
-                Ok(())
-            }
+            LiveEditMessage::NotSameDocument => Ok(()),
             LiveEditMessage::DocumentChange { resource_id } => {
                 info!("Peer changed document: {}", resource_id);
-
-                // Just set live editing inactive - the existing check_for_possible_disconnection
-                // will handle the timer on its own when needed
-                self.set_live_editing_inactive().await;
 
                 // Emit an event so the listener can remove this connection from active connections
                 self.event_emitter.emit(P2PEvent::DocumentChanged {
